update skill controllers

+++ controllers/skillController.js
@@ -7,7 +7,10 @@ export const getAllSkills = async(req, res) =>{
         res.status(200).json(response);
     })
     .catch((err)=>{
-        res.status(500).send("Something error happened! Can't load All skills");
+        res.status(500).json({            
+            message:"Something error happened! Can't load All skills",
+            error: err.message
+        });
     })
 }
 
@@ -40,13 +43,17 @@ export const updateSkill = async(req, res) =>{
         const id = req.params.id;
 
         if(!id){
-            return res.status(500).send(`Something error happened! Can't update skill document`);
+            return res.status(400).json({ message: "Skill ID is required for updating." });
+        }
+
+        if (!mongoose.Types.ObjectId.isValid(id)) {
+            return res.status(400).json({ message: "Invalid Skill ID format." });
         }
 
         const existingSkill = await skillModel.findById(id);
 
         if(!existingSkill){
-            return res.status(500).send(`Something error happened! Target skill document not exists`);
+            return res.status(500).json({ message: "Something error happened! Target skill document not found!" });
         }
 
         if (title) existingSkill.title = title;
@@ -71,9 +78,9 @@ export const deleteSkill = async (req, res)=>{
             return res.status(400).json({ message: "Skill ID is required for deletion." });
         }
 
-        // if (!mongoose.Types.ObjectId.isValid(id)) {
-        //     return res.status(400).json({ message: "Invalid Skill ID format." });
-        // }
+        if (!mongoose.Types.ObjectId.isValid(id)) {
+            return res.status(400).json({ message: "Invalid Skill ID format." });
+        }
 
         const deletedSkill = await skillModel.findByIdAndDelete(id);
 
@@ -81,7 +88,9 @@ export const deleteSkill = async (req, res)=>{
             return res.status(404).json({ message: "Skill not found or already deleted." });
         }
 
-        return res.send("Skill document deleted successfully")
+        return res.status(200).json({
+            message: "Skill document deleted successfully"
+        })
     }
     catch(err){
         return res.status(500).json({
